Fix various problems in Class Scheduling example for issue #76
Changes required to allow spawning the tasks using Tokio:
- Call rand::thread_rng() each time the random number generator is
  needed, instead of storing the rng or passing it as a parameter. The
  thread-local random number generator cannot be passed between
  threads, so this was preventing the futures returned by some of the
  async functions from being Send.
- Use transact_boxed instead of transact_boxed_local, so that the
  returned future is Send

After making the above changes, tokio::task::spawn can be used instead
of thread::spawn (with no need to call block_on).
The return value of tokio::task::spawn is a different kind of
JoinHandle, and the code needs to .await it instead of calling join().

Remove some comments that are not meaningful to the reader of the
example code.

Add a note explaining how the database will be modified by running the
code (especially since it does not use a directory).

Make the parameters of the simulation (number of students and number of
operations to attempt per student) const values for clarity.

Instead of using all 1,620 class names, and initializing the seat counts
to 100, use const values for number of classes and initial number of
seats, choosing a random selection of class names from the list of
possible combinations.

Use subspaces consistently throughout the code.
When getting the key range for querying the attends keys for the
student, use the .subspace method to get a Subspace struct for the
prefix. Add a comment explaining this because it seems
confusing/unintuitive (but I don't think the API provides a better way
of doing it).

Pass false for the snapshot parameter to the .get() method, and add
comments explaining why this is important (since this could be a major
gotcha for someone new to the API).

Make ditch_trx and signup_trx both return bool values indicating
whether the operation was performed. If the user was not in the class,
or was already in the class, return false. Obviously it could return an
Err here to cancel the transaction, but returning success values other
than () adds variety to the example code.

Make switch_classes() handle the cases where the student was not in the
old class, or was already in the new class, canceling the transaction
in each case. Define two new error codes for the two cases.

In perform_op(), check the return values of ditch() and signup(), and
update my_classes only if the operation succeeds. A similar change is
not required for the switch case, because that will return an error
result.

Log the operations that were successfully performed, to make the
simulation more interesting. (The code already has commented-out
printlns inside the transactions, but that logs operations even if
they failed, or logs them multiple times if the transaction is
repeated, which produces very cluttered and confusing output.)

Add a comment explaining why get_available_classes() is called when it
is, since that is one of the more confusing parts of the example.

Prevent panics when trying to choose values from my_classes or
available_classes and the vectors are empty.

Use get_ranges_keyvalues() and get_ranges() instead of get_range(),
since get_range() is not recommended for use by application code and
has a confusing parameter (see #71 / #72).
In get_available_classes() and signup_trx(), use get_ranges_keyvalues()
the easy way, using .try_collect() to first get all results as a
vector.
In run_sim(), provide a commented-out example of iterating through the
Stream returned by get_ranges_keyvalues(), but use get_ranges() here
instead.