Modify class-scheduling.rs to use `Database::run`, addressing #59, and improve README.md for the examples folder
Change the `Result<T>` type (returned by various functions) to use
`FdbBindingError` instead of the custom `Error` type.
`Error` no longer needs to wrap `FdbError`, so the `Internal` variant
can be removed.

Implement the `std::error::Error` trait for the custom Error type.
This is needed because `FdbBindingError` can only wrap custom errors
if they implement `std::error::Error`.
Requires implementations of `Display` and `Debug`.

Add a `wrap_custom_error` function, because wrapping an error is a
bit complicated, and it's needed in four places.

Change the type of the `txn` parameter to `RetryableTransaction` for
`ditch_trx`, `signup_trx`, and `switch_classes_body`.
Change the string parameters to type `String` to avoid lifetime
errors with `&str` references due to the combination of closures and
`async` functions.

Handle the `FdbBindingError` returned by perform_op.
The error handingcode now differentiates between custom errors (which
require calling `get_available_classes` to update the cache) and
other errors, which are fatal.

Add comment explaining that class-scheduling.rs is not compatible with v0.7.0

Also add descriptions of the examples to README.md