feat(tuple) add function to pack a subspace key with a versionstamp
Adding a convenience method to support versionstamped keys within a
subpsace.

This supports any level of the subspace having incomplete versionstamps
which are merged when packing new keys and subspaces.

It keeps track of the current versionstamp alongside the prefix since
we need to know how to adjust the offsets and how many incomplete
versionstamps we have.